Require Python 3 tests to pass in Django 1.7
South never really supported Python 3, but 0.8 was good enough to not
cause any immediately visible issues. South 1.0 now fails hard when
running Python 3.

This leaves us with the option to support both South 0.8 and 1.0. But as
we're just introducing Python 3 support and South is end-of-life anyway,
I've opted to require using Django 1.7 if running Python 3. This
combination is what new deployments will likely choose anyway.